chore(website): optimize images, move logo to assets folder
