+++ libs/website/index-page/src/lib/hero-block/hero-block.tsx
@@ -1,7 +1,7 @@
 import { Text, Button, HeroHeading } from '@haqq/website/ui-kit';
 import Image from 'next/image';
-import sunriseImageData from '../../assets/sunrise.png';
-import halfEclipseImageData from '../../assets/half-eclipse.png';
+import sunriseImageData from '../../assets/images/sunrise.png';
+import halfEclipseImageData from '../../assets/images/half-eclipse.png';
 
 function SunriseBackground() {
   return (
