+++ libs/website/ui-kit/src/lib/header/header.tsx
@@ -4,6 +4,7 @@ import clsx from 'clsx';
 import Link from 'next/link';
 import { Button } from '../button/button';
 import { BurgerButton } from '../burger-button/burger-button';
+import logoImageData from '../../assets/images/logo.svg';
 
 function HeaderNavLink({
   href,
@@ -30,7 +31,7 @@ export function Header() {
       <div className="w-full flex flex-row items-center h-full pr-[16px] sm:pr-[64px] lg:pr-[80px] mx-auto">
         <div className="w-[48px] sm:w-[64px] lg:w-[80px] h-full flex items-center justify-center border-r border-haqq-border">
           <div className="relative w-[26px] h-[26px] sm:w-[32px] sm:h-[32px]">
-            <Image src="/logo.svg" alt="HAQQ" fill />
+            <Image src={logoImageData.src} alt="HAQQ" fill />
           </div>
         </div>
         <div className="ml-[12px] sm:ml-[20px] lg:ml-[32px] font-serif font-medium text-[20px] sm:text-[24px] leading-none">
